+++ pkg/core/hooks/conn/factory.go
@@ -3,11 +3,7 @@
 package conn
 
 import (
-	"bytes"
 	"context"
-	"io"
-	"mime/multipart"
-	"strings"
 	"sync"
 	"time"
 
@@ -92,47 +88,4 @@ func (factory *Factory) GetOrCreate(connectionID ID) *Tracker {
 		return factory.connections[connectionID]
 	}
 	return tracker
-}
-
-func extractFormData(logger *zap.Logger, body []byte, contentType string) []models.FormData {
-	boundary := ""
-	if strings.HasPrefix(contentType, "multipart/form-data") {
-		parts := strings.Split(contentType, "boundary=")
-		if len(parts) > 1 {
-			boundary = strings.TrimSpace(parts[1])
-		} else {
-			utils.LogError(logger, nil, "Invalid multipart/form-data content type")
-			return nil
-		}
-	}
-	reader := multipart.NewReader(bytes.NewReader(body), boundary)
-	var formData []models.FormData
-
-	for {
-		part, err := reader.NextPart()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			utils.LogError(logger, err, "Error reading part")
-			continue
-		}
-		key := part.FormName()
-		if key == "" {
-			continue
-		}
-
-		value, err := io.ReadAll(part)
-		if err != nil {
-			utils.LogError(logger, err, "Error reading part value")
-			continue
-		}
-
-		formData = append(formData, models.FormData{
-			Key:    key,
-			Values: []string{string(value)},
-		})
-	}
-
-	return formData
 }
