+++ pkg/core/proxy/integrations/http/util.go
@@ -1,5 +1,3 @@
-//go:build linux
-
 package http
 
 import (
