+++ pkg/core/proxy/util/util.go
@@ -1,5 +1,3 @@
-//go:build linux
-
 // Package util provides utility functions for the proxy package.
 package util
